ubah int length

+++ database/migrations/2019_11_23_091718_add_columns_to_attendances.php
@@ -28,7 +28,7 @@ class AddColumnsToAttendances extends Migration
             $table->unsignedTinyInteger('evaluation');
             $table->unsignedTinyInteger('additional_type');
             $table->time('boundary');
-            $table->unsignedTinyInteger('additional_minutes')
+            $table->unsignedInteger('additional_minutes')
                 ->nullable();
         });
     }
